requirements.txt
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
redis==5.0.1
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.1
ccxt==4.1.10
web3==6.13.2
python-binance==1.0.19
python-telegram-bot==20.7
prometheus-client==0.19.0
sentry-sdk==1.39.1
python-jose==3.3.0
aiohttp==3.9.1
python-socketio==5.12.0
python-telegram-bot==20.7
python-dotenv==1.0.0
